feat: prettify SGID login button on app
