+++ src/features/sign-in/components/index.ts
@@ -4,3 +4,5 @@ export * from './LoginImageSvgr'
 export * from './Emailnput'
 export * from './VerificationInput'
 export * from './SignInForm'
+
+export * from './SgidLoginButton'
